refactor :  코드리뷰 반영 및 폴더구조 변경
