+++ src/main/java/roomescape/RoomescapeController.java
@@ -6,7 +6,12 @@ import org.springframework.web.bind.annotation.GetMapping;
 @Controller
 public class RoomescapeController {
     @GetMapping("/")
-    public String home(){
+    public String home() {
         return "home";
+    }
+
+    @GetMapping("/reservation")
+    public String reservation() {
+        return "reservation";
     }
 }
