+++ src/main/java/roomescape/reservation/api/ReservationController.java
@@ -1,43 +1,30 @@
-package roomescape;
+package roomescape.reservation.api;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import roomescape.reservation.domain.Reservation;
 
 import java.net.URI;
-import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicLong;
 
-@Controller
+@RestController
+@RequestMapping("/reservations")
 public class ReservationController {
     private List<Reservation> reservations = new ArrayList<>();
     private AtomicLong index = new AtomicLong(1);
 
-    @GetMapping("/reservation")
-    public String reservation(Model model) {
-        if (reservations.isEmpty()) {
-            reservations.add(new Reservation(index.getAndIncrement(), "브라운", LocalDate.of(2023, 1, 1), LocalTime.of(10, 0, 0)));
-            reservations.add(new Reservation(index.getAndIncrement(), "브라운", LocalDate.of(2023, 1, 2), LocalTime.of(11, 0, 0)));
-            reservations.add(new Reservation(index.getAndIncrement(), "브라운", LocalDate.of(2023, 1, 3), LocalTime.of(12, 0, 0)));
-        }
-
-        return "reservation";
-    }
-
-    @GetMapping("/reservations")
+    @GetMapping
     public ResponseEntity<List<Reservation>> read() {
         return ResponseEntity.status(HttpStatus.OK).body(reservations);
     }
 
-    @PostMapping("/reservations")
+    @PostMapping
     public ResponseEntity<Reservation> create(@RequestBody Reservation reservation){
-        if (reservation.getName() == null || reservation.getDate() == null || reservation.getTime() == null) {
+        if (reservation.getName() == null || reservation.getDate() == null || reservation.getTime().isEmpty()) {
             throw new NotFoundReservationException("빈 값이 존재합니다!");
         }
         if (reservations.stream().anyMatch(r -> Objects.equals(r.getName(), reservation.getName()))) {
@@ -52,7 +39,7 @@ public class ReservationController {
                 .body(newReservation);
     }
 
-    @DeleteMapping("/reservations/{id}")
+    @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
         Reservation reservation = reservations.stream()
                 .filter(it -> Objects.equals(it.getId(), id))
